swarm/pss: Fixed missing protoCtrl on go build tag < 1.8
