+++ swarm/pss/handshake_test.go
@@ -1,4 +1,4 @@
-// +build go1.8
+// +build foo
 
 package pss
